[Fix] AVX, AVX2: Clean-up failing unit tests.
Former-commit-id: e46ac602094e6485649a67075a805c2b7a1bba23

+++ plugins/avx/float/UMESimdVecFloat64_2.h
@@ -1180,16 +1180,16 @@ namespace SIMD {
             return SIMDVec_f(t0, t1);
         }
         // TRUNC
-        inline SIMDVec_i<int32_t, 2> trunc() const {
-            int32_t t0 = (int32_t)mVec[0];
-            int32_t t1 = (int32_t)mVec[1];
-            return SIMDVec_i<int32_t, 2>(t0, t1);
+        inline SIMDVec_i<int64_t, 2> trunc() const {
+            int64_t t0 = (int64_t)mVec[0];
+            int64_t t1 = (int64_t)mVec[1];
+            return SIMDVec_i<int64_t, 2>(t0, t1);
         }
         // MTRUNC
-        inline SIMDVec_i<int32_t, 2> trunc(SIMDVecMask<2> const & mask) const {
-            int32_t t0 = mask.mMask[0] ? (int32_t)mVec[0] : 0;
-            int32_t t1 = mask.mMask[1] ? (int32_t)mVec[1] : 0;
-            return SIMDVec_i<int32_t, 2>(t0, t1);
+        inline SIMDVec_i<int64_t, 2> trunc(SIMDVecMask<2> const & mask) const {
+            int64_t t0 = mask.mMask[0] ? (int64_t)mVec[0] : 0;
+            int64_t t1 = mask.mMask[1] ? (int64_t)mVec[1] : 0;
+            return SIMDVec_i<int64_t, 2>(t0, t1);
         }
         // FLOOR
         // MFLOOR
